requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-modeltranslation==0.19.10
idna==3.11
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
telebot==0.0.5
urllib3==2.5.0

# File processing libraries
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==11.0.0
python-magic==0.4.27
psycopg2-binary